Create README.md
This is a Classification Project created from a dataset obtained from UCI Repository. The dataset consists of 6 features including buying, maintenance, number of persons that can fit in the car, number of doors of the car, safety and the class of the car. The project aims to classify the vehicle with the corresponding characteristics into 2 categories- acceptable or unacceptable. The project has been pre-processed and trained with 6 models and a real time prediction has also been performed.